Fixed design issues
