+++ src/elements/Input.scss
@@ -118,7 +118,7 @@
   input,
   textarea,
   select {
-    color: colors.$gray-300;
+    color: colors.$white;
     background-color: colors.$gray-600;
     border: 1px solid colors.$gray-500;
