+++ src/views/web3-tools/Web3Converter.scss
@@ -15,8 +15,13 @@
     }
 
     input:invalid {
-      border: 1px solid colors.$error-700;
-      box-shadow: 0px 1px 2px rgb(16 24 40 / 5%), 0px 0px 0px 4px colors.$error-900 !important;
+      border: 1px solid colors.$error-600;
+      box-shadow: 0px 1px 2px rgb(16 24 40 / 5%), 0px 0px 0px 4px colors.$error-300 !important;
+
+      .dark & {
+        border: 1px solid colors.$error-700;
+        box-shadow: 0px 1px 2px rgb(16 24 40 / 5%), 0px 0px 0px 4px colors.$error-900 !important;
+      }
     }
 
     textarea {
